refactoring signer mod.
