+++ integration/src/tests/extension_omni_lock/build_transfer_transaction_ckb.rs
@@ -4,7 +4,7 @@ use crate::utils::address::generate_rand_secp_address_pk_pair;
 use crate::utils::address::omni_lock::{
     prepare_omni_ethereum_address_with_capacity, prepare_omni_secp_address_with_capacity,
 };
-use crate::utils::instruction::{dump_data, send_transaction_to_ckb};
+use crate::utils::instruction::send_transaction_to_ckb;
 use crate::utils::rpc_client::MercuryRpcClient;
 use crate::utils::signer::sign_transaction;
 use core_rpc_types::{
@@ -59,6 +59,70 @@ fn test_omni_secp_transfer_ckb() {
     let tx = mercury_client.build_transfer_transaction(payload).unwrap();
     let tx = sign_transaction(tx, &pks).unwrap();
     // dump_data(&tx, "tx.json").unwrap();
+
+    // send tx to ckb node
+    let _tx_hash = send_transaction_to_ckb(tx).unwrap();
+
+    // get balance of to address
+    let mut asset_infos = HashSet::new();
+    asset_infos.insert(AssetInfo::new_ckb());
+    let payload = GetBalancePayload {
+        item: JsonItem::Address(to_address.to_string()),
+        asset_infos,
+        extra: None,
+        tip_block_number: None,
+    };
+    let to_balance = mercury_client.get_balance(payload).unwrap();
+
+    assert_eq!(to_balance.balances.len(), 1);
+    assert_eq!(to_balance.balances[0].asset_info.asset_type, AssetType::CKB);
+    assert_eq!(to_balance.balances[0].free, 100_0000_0000u128.into());
+}
+
+inventory::submit!(IntegrationTest {
+    name: "test_omni_eth_transfer_ckb",
+    test_fn: test_omni_eth_transfer_ckb
+});
+fn test_omni_eth_transfer_ckb() {
+    let (identity, address, address_pk, _out_point) =
+        prepare_omni_ethereum_address_with_capacity(300_0000_0000).expect("prepare ckb");
+    let pks = vec![address_pk];
+    let mercury_client = MercuryRpcClient::new(MERCURY_URI.to_string());
+
+    let (to_address, _to_pk) = generate_rand_secp_address_pk_pair();
+
+    let mut asset_infos = HashSet::new();
+    asset_infos.insert(AssetInfo::new_ckb());
+    let balance_payload = GetBalancePayload {
+        item: JsonItem::Identity(identity.encode()),
+        asset_infos,
+        extra: None,
+        tip_block_number: None,
+    };
+    let balance = mercury_client.get_balance(balance_payload).unwrap();
+    assert_eq!(balance.balances.len(), 1);
+    assert_eq!(balance.balances[0].asset_info.asset_type, AssetType::CKB);
+    assert_eq!(balance.balances[0].occupied, 0u128.into());
+    assert_eq!(balance.balances[0].free, 300_0000_0000u128.into());
+    assert_eq!(balance.balances[0].frozen, 0u128.into());
+
+    let payload = TransferPayload {
+        asset_info: AssetInfo::new_ckb(),
+        from: vec![JsonItem::Address(address.to_string())],
+        to: vec![ToInfo {
+            address: to_address.to_string(),
+            amount: 100_0000_0000u128.into(),
+        }],
+        output_capacity_provider: Some(OutputCapacityProvider::From),
+        pay_fee: None,
+        fee_rate: None,
+        since: None,
+    };
+
+    // build tx
+    let mercury_client = MercuryRpcClient::new(MERCURY_URI.to_string());
+    let tx = mercury_client.build_transfer_transaction(payload).unwrap();
+    let tx = sign_transaction(tx, &pks).unwrap();
 
     // send tx to ckb node
     let _tx_hash = send_transaction_to_ckb(tx).unwrap();
