--- integration/src/utils/signer.rs
@@ -1,349 +1,0 @@
-use crate::const_definition::{
-    ANYONE_CAN_PAY_DEVNET_TYPE_HASH, CHEQUE_DEVNET_TYPE_HASH, OMNI_LOCK_DEVNET_TYPE_HASH,
-    PW_LOCK_DEVNET_TYPE_HASH, SIGHASH_TYPE_HASH,
-};
-use crate::utils::address::pubkey_to_eth_address;
-
-use anyhow::Result;
-use ckb_crypto::secp::Privkey;
-use ckb_hash::blake2b_256;
-use ckb_hash::new_blake2b;
-use ckb_jsonrpc_types::{Script, Transaction};
-use ckb_types::{
-    bytes::Bytes,
-    packed::{self, BytesOpt},
-    prelude::*,
-    H160, H256,
-};
-use core_rpc_types::{Identity, IdentityFlag, ScriptGroupType, TransactionCompletionResponse};
-use core_storage::OmniLockWitnessLock;
-use crypto::digest::Digest;
-use crypto::sha3::Sha3;
-use secp256k1::{self, PublicKey, Secp256k1, SecretKey};
-
-use std::convert::From;
-use std::str::FromStr;
-
-pub fn sign_transaction(
-    transaction: TransactionCompletionResponse,
-    pks: &[H256],
-) -> Result<Transaction> {
-    let script_groups = transaction.script_groups;
-    let tx_hash = transaction.tx_view.hash.0;
-    let tx: packed::Transaction = transaction.tx_view.inner.into();
-    let mut witnesses: Vec<packed::Bytes> = tx.witnesses().into_iter().collect();
-    for script_group in script_groups {
-        if script_group.group_type == ScriptGroupType::Type {
-            continue;
-        }
-        let pk = if let Some(pk) = get_right_pk(pks, &script_group.script) {
-            pk
-        } else {
-            continue;
-        };
-        let init_witness_idx: u32 = script_group.input_indices[0].into();
-        if witnesses[init_witness_idx as usize].to_string() == packed::Bytes::default().to_string()
-        {
-            continue;
-        }
-        if script_group.script.code_hash == SIGHASH_TYPE_HASH
-            || script_group.script.code_hash == ANYONE_CAN_PAY_DEVNET_TYPE_HASH
-            || script_group.script.code_hash == CHEQUE_DEVNET_TYPE_HASH
-        {
-            let init_witness = if witnesses[init_witness_idx as usize].is_empty() {
-                packed::WitnessArgs::default()
-            } else {
-                packed::WitnessArgs::from_slice(
-                    witnesses[init_witness_idx as usize].raw_data().as_ref(),
-                )
-                .map_err(anyhow::Error::new)?
-            };
-
-            let init_witness = init_witness
-                .as_builder()
-                .lock(Some(Bytes::from(vec![0u8; 65])).pack())
-                .build();
-            let mut blake2b = new_blake2b();
-            blake2b.update(&tx_hash);
-            blake2b.update(&(init_witness.as_bytes().len() as u64).to_le_bytes());
-            blake2b.update(&init_witness.as_bytes());
-            for idx in &script_group.input_indices[1..] {
-                let idx: u32 = (*idx).into();
-                let other_witness = witnesses[idx as usize].raw_data();
-                blake2b.update(&(other_witness.len() as u64).to_le_bytes());
-                blake2b.update(&other_witness);
-            }
-            for other_witness in witnesses.iter().skip(tx.raw().inputs().len()) {
-                let other_witness = other_witness.raw_data();
-                blake2b.update(&(other_witness.len() as u64).to_le_bytes());
-                blake2b.update(&other_witness);
-            }
-            let mut message = [0u8; 32];
-            blake2b.finalize(&mut message);
-            let message = H256::from(message);
-
-            let privkey = Privkey::from_slice(pk.as_bytes());
-            let sig = privkey.sign_recoverable(&message).expect("sign");
-            witnesses[init_witness_idx as usize] = init_witness
-                .as_builder()
-                .lock(Some(Bytes::from(sig.serialize())).pack())
-                .build()
-                .as_bytes()
-                .pack();
-        } else if script_group.script.code_hash == PW_LOCK_DEVNET_TYPE_HASH {
-            let init_witness = if witnesses[init_witness_idx as usize].is_empty() {
-                packed::WitnessArgs::default()
-            } else {
-                packed::WitnessArgs::from_slice(
-                    witnesses[init_witness_idx as usize].raw_data().as_ref(),
-                )
-                .map_err(anyhow::Error::new)?
-            };
-
-            let init_witness = init_witness
-                .as_builder()
-                .lock(Some(Bytes::from(vec![0u8; 65])).pack())
-                .build();
-            let mut hasher = Sha3::keccak256();
-            hasher.input(&tx_hash);
-            hasher.input(&(init_witness.as_bytes().len() as u64).to_le_bytes());
-            hasher.input(&init_witness.as_bytes());
-            for idx in &script_group.input_indices[1..] {
-                let idx: u32 = (*idx).into();
-                let other_witness = witnesses[idx as usize].raw_data();
-                hasher.input(&(other_witness.len() as u64).to_le_bytes());
-                hasher.input(&other_witness);
-            }
-            for other_witness in witnesses.iter().skip(tx.raw().inputs().len()) {
-                let other_witness = other_witness.raw_data();
-                hasher.input(&(other_witness.len() as u64).to_le_bytes());
-                hasher.input(&other_witness);
-            }
-            let mut message = [0u8; 32];
-            hasher.result(&mut message);
-
-            hash_personal_message(&mut message);
-
-            let message = H256::from(message);
-
-            let privkey = Privkey::from_slice(pk.as_bytes());
-            let sig = privkey.sign_recoverable(&message).expect("sign");
-            witnesses[init_witness_idx as usize] = init_witness
-                .as_builder()
-                .lock(Some(Bytes::from(sig.serialize())).pack())
-                .build()
-                .as_bytes()
-                .pack();
-        } else if script_group.script.code_hash == OMNI_LOCK_DEVNET_TYPE_HASH {
-            let ident = omni_script_to_identity(&script_group.script.into()).unwrap();
-            let (flag, _pubkey_hash) = ident.parse().unwrap();
-            match flag {
-                IdentityFlag::Ckb => {
-                    let init_witness = if witnesses[init_witness_idx as usize].is_empty() {
-                        packed::WitnessArgs::default()
-                    } else {
-                        packed::WitnessArgs::from_slice(
-                            witnesses[init_witness_idx as usize].raw_data().as_ref(),
-                        )
-                        .map_err(anyhow::Error::new)?
-                    };
-
-                    let witness_lock = OmniLockWitnessLock::new_builder()
-                        .signature(Some(Bytes::from(vec![0u8; 65])).pack())
-                        .build();
-                    let len = witness_lock.as_bytes().len();
-                    let zero_lock = Bytes::from(vec![0u8; len]);
-
-                    let init_witness = init_witness
-                        .as_builder()
-                        .lock(Some(zero_lock).pack())
-                        .build();
-                    let mut blake2b = new_blake2b();
-                    blake2b.update(&tx_hash);
-                    blake2b.update(&(init_witness.as_bytes().len() as u64).to_le_bytes());
-                    blake2b.update(&init_witness.as_bytes());
-                    for idx in &script_group.input_indices[1..] {
-                        let idx: u32 = (*idx).into();
-                        let other_witness = witnesses[idx as usize].raw_data();
-                        blake2b.update(&(other_witness.len() as u64).to_le_bytes());
-                        blake2b.update(&other_witness);
-                    }
-                    for other_witness in witnesses.iter().skip(tx.raw().inputs().len()) {
-                        let other_witness = other_witness.raw_data();
-                        blake2b.update(&(other_witness.len() as u64).to_le_bytes());
-                        blake2b.update(&other_witness);
-                    }
-                    let mut message = [0u8; 32];
-                    blake2b.finalize(&mut message);
-                    let message = H256::from(message);
-
-                    let privkey = Privkey::from_slice(pk.as_bytes());
-                    let sig = privkey.sign_recoverable(&message).expect("sign");
-
-                    // Put signature into witness
-                    let current_witness = if witnesses[init_witness_idx as usize].is_empty() {
-                        packed::WitnessArgs::default()
-                    } else {
-                        packed::WitnessArgs::from_slice(
-                            witnesses[init_witness_idx as usize].raw_data().as_ref(),
-                        )
-                        .map_err(anyhow::Error::new)?
-                    };
-
-                    let orig_lock = current_witness.lock();
-                    let lock_field = orig_lock.to_opt().map(|data| data.raw_data());
-                    let omnilock_witnesslock = if let Some(lock_field) = lock_field {
-                        OmniLockWitnessLock::from_slice(lock_field.as_ref())?
-                    } else {
-                        OmniLockWitnessLock::default()
-                    };
-                    let omnilock_witnesslock = omnilock_witnesslock
-                        .as_builder()
-                        .signature(Some(Bytes::from(sig.serialize())).pack())
-                        .build();
-                    let witness_lock = Some(omnilock_witnesslock.as_bytes()).pack();
-                    witnesses[init_witness_idx as usize] = current_witness
-                        .as_builder()
-                        .lock(witness_lock)
-                        .build()
-                        .as_bytes()
-                        .pack();
-                }
-                IdentityFlag::Ethereum => {}
-                _ => {}
-            }
-        } else {
-            todo!()
-        }
-    }
-
-    let tx = tx
-        .as_advanced_builder()
-        .set_witnesses(witnesses)
-        .build()
-        .data();
-    Ok(tx.into())
-}
-
-/// Build proper witness lock
-pub fn build_witness_lock(orig_lock: BytesOpt, signature: Bytes) -> Result<Bytes> {
-    let lock_field = orig_lock.to_opt().map(|data| data.raw_data());
-    let omnilock_witnesslock = if let Some(lock_field) = lock_field {
-        OmniLockWitnessLock::from_slice(lock_field.as_ref())?
-    } else {
-        OmniLockWitnessLock::default()
-    };
-
-    Ok(omnilock_witnesslock
-        .as_builder()
-        .signature(Some(signature).pack())
-        .build()
-        .as_bytes())
-}
-
-pub fn sign_transaction_for_cheque_of_sender(
-    mut transaction: TransactionCompletionResponse,
-    pk: &H256,
-    cheque_input_indexes: Vec<usize>,
-) -> Result<Transaction> {
-    for index in cheque_input_indexes {
-        let since = &mut transaction.tx_view.inner.inputs[index].since;
-        *since = {
-            // when sender withdraw, cheque cell since must be hardcoded as 0xA000000000000006
-            11529215046068469766u64.into()
-        };
-    }
-    let tx: packed::Transaction = transaction.tx_view.inner.into();
-    let tx_view = tx.as_advanced_builder().build();
-    transaction.tx_view = tx_view.into();
-    sign_transaction(transaction, &[pk.to_owned()])
-}
-
-pub fn get_uncompressed_pubkey_from_pk(pk: &str) -> String {
-    let secret_key = SecretKey::from_str(pk).expect("get SecretKey");
-    let secp256k1: Secp256k1<secp256k1::All> = secp256k1::Secp256k1::new();
-    let pubkey = PublicKey::from_secret_key(&secp256k1, &secret_key);
-    hex::encode(pubkey.serialize_uncompressed())
-}
-
-fn pubkey_to_secp_lock_arg(pubkey_uncompressed: &str) -> Bytes {
-    let pubkey = secp256k1::PublicKey::from_str(pubkey_uncompressed).unwrap();
-    let pubkey_compressed = &pubkey.serialize()[..];
-
-    assert_eq!(33, pubkey_compressed.len());
-
-    let pubkey_hash = blake2b_256(pubkey_compressed);
-
-    assert_eq!(32, pubkey_hash.len());
-
-    let pubkey_hash = &pubkey_hash[0..20];
-    let pubkey_hash =
-        H160::from_slice(pubkey_hash).expect("Generate hash(H160) from pubkey failed");
-    Bytes::from(pubkey_hash.as_bytes().to_vec())
-}
-
-fn get_secp_lock_arg(pk: &H256) -> Bytes {
-    let pubkey = get_uncompressed_pubkey_from_pk(&pk.to_string());
-    pubkey_to_secp_lock_arg(&pubkey)
-}
-
-fn get_pw_lock_arg(pk: &H256) -> Bytes {
-    let pubkey = get_uncompressed_pubkey_from_pk(&pk.to_string());
-    let args = pubkey_to_eth_address(&pubkey);
-    let args = H160::from_str(&args).expect("get args");
-    Bytes::copy_from_slice(args.as_bytes())
-}
-
-fn get_omni_secp_arg(pk: &H256) -> Bytes {
-    let pubkey = get_uncompressed_pubkey_from_pk(&pk.to_string());
-    let args = pubkey_to_secp_lock_arg(&pubkey).to_vec();
-    let mut auth = vec![0u8];
-    auth.extend_from_slice(&args);
-    Bytes::copy_from_slice(&auth)
-}
-
-fn get_omni_eth_arg(pk: &H256) -> Bytes {
-    let pubkey = get_uncompressed_pubkey_from_pk(&pk.to_string());
-    let args = pubkey_to_eth_address(&pubkey);
-    let args = H160::from_str(&args).expect("get args");
-    let mut auth = vec![1u8];
-    auth.extend_from_slice(args.as_bytes());
-    Bytes::copy_from_slice(&auth)
-}
-
-fn get_right_pk<'a>(pks: &'a [H256], script: &Script) -> Option<&'a H256> {
-    if script.code_hash == CHEQUE_DEVNET_TYPE_HASH {
-        return Some(&pks[0]);
-    }
-    let args = script.args.clone().into_bytes();
-    for pk in pks {
-        if get_secp_lock_arg(pk) == args || get_pw_lock_arg(pk) == args {
-            return Some(pk);
-        }
-        if args.len() > 21 {
-            if get_omni_secp_arg(pk).to_vec() == args.to_vec()[0..21]
-                || get_omni_eth_arg(pk).to_vec() == args.to_vec()[0..21]
-            {
-                return Some(pk);
-            }
-        }
-    }
-    None
-}
-
-fn hash_personal_message(message: &mut [u8; 32]) {
-    let prefix = format!("\x19Ethereum Signed Message:\n{}", message.len())
-        .as_bytes()
-        .to_vec();
-    let new = [prefix, message.to_vec()].concat();
-
-    let mut hasher = Sha3::keccak256();
-    hasher.input(&new);
-    hasher.result(message);
-}
-
-fn omni_script_to_identity(script: &packed::Script) -> Option<Identity> {
-    let flag = script.args().as_slice()[4..5].to_vec()[0].try_into().ok()?;
-    let hash = H160::from_slice(&script.args().as_slice()[5..25]).ok()?;
-    Some(Identity::new(flag, hash))
-}
